Adjust the codemark form at the bottom of the viewport
https://trello.com/c/XyLA3yDb/2407-display-of-codemarks-prs-compose-box-is-off-near-bottom-of-file

This is still a bit wonky I think because the form will persist visibly
at the bottom of the viewport until the corresponding line in the editor
is no longer visible. But when the form is at the top of the viewport,
it can be partially visible as the user scrolls.