Use `AgentConnection` instead of an instance to a running agent
Fixes https://trello.com/c/B0BnIBhi/2841-not-receiving-real-time-new-codemarks-in-activity-feed-in-atom

Subscription to notifications from the agent weren't happening when
a log out and log in occurred within the same window without a reload.

This refactoring uses a single AgentConnection instance that can
handle starting/stopping the running agent while allowing consumers to
keep the same reference to an agent even when the running agent is
disposed and replaced.