Clean up feature flags
I created a `isFeatureEnabled` selector to do the capabilities check.
This is useful because it creates a single point to hardcode for certain strings or be more complex about the check.
The <FeatureFlag/> uses the selector and can be used in rendering contexts if the consumer doesn't need to know the state of the flag.